requirements.txt
requests
pandas
python-dotenv
browser-cookie3
